==> 25m/01_data_preprocessing_(added_imdb_context)/03_edit_movie_context.ipynb <==
# Edit movie context

import pandas as pd
import json

# depth of this file in the project
file_depth = '../..'

with open(file_depth + '/config/data_25m_config.json') as config_file:
    config = json.load(config_file)

new_context_path = config['new_context_path']
movies_with_context = pd.read_csv(file_depth + new_context_path + 'backup/movies_with_context.csv', encoding="UTF-8")
movies_with_context

### Remove writers column

movies_with_context.drop('writers', axis=1, inplace=True)
movies_with_context

### In columns director and actor, keep only the first ID

def keep_first_id(cell):
    if ',' in cell:
        return cell.split(',')[0].strip()
    else:
        return cell.strip()
    
movies_with_context['directors'] = movies_with_context['directors'].apply(keep_first_id)
movies_with_context['actor'] = movies_with_context['actor'].apply(keep_first_id)
movies_with_context

### One-hot encode movie genres

genres = movies_with_context['genres'].str.lower().str.get_dummies('|')
genres.columns = genres.columns.str.capitalize()
genres = genres.add_prefix('genre')
genres

genre_columns = [col for col in genres.columns if col.startswith('genre')]
column_names = [col.replace('genre', '') for col in genre_columns]
print(f"All {len(column_names)} unique movie genres/categories: ", column_names)

movies_with_context = pd.concat([movies_with_context, genres], axis=1)
movies_with_context = movies_with_context.drop(columns=['genres'])
movies_with_context

Replace '\N' values in column directors with 'no_director'

print('Number of \"\\N\" values in column directors: ', len(movies_with_context[movies_with_context['directors'] == '\\N']))
movies_with_context['directors'] = movies_with_context['directors'].replace('\\N', 'no_director')
print('Number of \"\\N\" values in column directors: ', len(movies_with_context[movies_with_context['directors'] == '\\N']))

Replace '\N' values in column runtimeMinutes with 0 and convert the column to int

print('Number of \"\\N\" values in column runtimeMinutes: ', len(movies_with_context[movies_with_context['runtimeMinutes'] == '\\N']))
movies_with_context['runtimeMinutes'] = movies_with_context['runtimeMinutes'].replace('\\N', '0')
print('Number of \"\\N\" values in column runtimeMinutes: ', len(movies_with_context[movies_with_context['directors'] == '\\N']))

movies_with_context['runtimeMinutes'] = movies_with_context['runtimeMinutes'].astype(int)

---

movies_with_context.to_csv(file_depth + new_context_path + 'movies_with_context.csv', index=False)
