==> 25m/01_data_preprocessing_(added_imdb_context)/06_split_data.ipynb <==
from sklearn.preprocessing import LabelEncoder, MaxAbsScaler
from sklearn.model_selection import train_test_split
from joblib import dump
import pandas as pd
import numpy as np
import json

# depth of this file in the project
file_depth = '../..'

with open(file_depth + '/config/data_25m_config.json') as config_file:
    config = json.load(config_file)

new_context_data = config['new_context_data']
added_imdb_context = pd.read_csv(file_depth + new_context_data)
added_imdb_context

added_imdb_context.info()

all_non_null = []
all_non_null.append((added_imdb_context['runtimeMinutes'] == '\\N').any())
all_non_null.append((added_imdb_context['holiday'] == '\\N').any())
all_non_null.append((added_imdb_context['titleType'] == '\\N').any())
all_non_null.append((added_imdb_context['directors'] == '\\N').any())
all_non_null.append((added_imdb_context['actor'] == '\\N').any())

if any(item == True for item in all_non_null):
    print("There are null value in the columns with type object.")
else:
    print("There are no null value in the columns with type object.")

target_data = config['target_data']
ratings_target = pd.read_csv(file_depth + target_data)
ratings_target

added_imdb_context['ratingId'] = added_imdb_context.index
added_imdb_context = added_imdb_context.merge(ratings_target, on='ratingId', how='left')
del ratings_target
added_imdb_context = added_imdb_context.drop(['ratingId'], axis=1)
added_imdb_context

# Label encode categorical columns
categorical_columns = ['holiday', 'titleType', 'directors', 'actor']
label_encoders = {}

for column in categorical_columns:
    if column in added_imdb_context.columns.to_list():
        print(f"Column {column} is in context_data")
        label_encoder = LabelEncoder()
        added_imdb_context[column] = label_encoder.fit_transform(added_imdb_context[column])
        label_encoders[column] = label_encoder

added_imdb_context

train, temp = train_test_split(added_imdb_context, test_size=0.2, random_state=42)
del added_imdb_context
val, test = train_test_split(temp, test_size=0.5, random_state=42)
del temp

----

train_target_ratings = train['rating'].values
val_target_ratings = val['rating'].values
test_target_ratings = test['rating'].values

train_data = train.drop(['rating'], axis=1)
val_data = val.drop(['rating'], axis=1)
test_data = test.drop(['rating'], axis=1)

# # Standardize columns

scaler = MaxAbsScaler()
# target_scaler = MaxAbsScaler()

train_data = scaler.fit_transform(train_data)
val_data = scaler.fit_transform(val_data)
test_data = scaler.fit_transform(test_data)

# train_target_ratings = target_scaler.fit_transform(train_target_ratings.reshape(-1, 1))
# val_target_ratings = target_scaler.fit_transform(val_target_ratings.reshape(-1, 1))
# test_target_ratings = target_scaler.fit_transform(test_target_ratings.reshape(-1, 1))

# Save the data to disk

training_data_path = config['training_data']
data_group = 'data_25m/'
folder = 'added_imdb_context/'

dump(scaler, file_depth + training_data_path + f"{data_group}{folder}25m_added_imdb_context_scaler.pkl")
# dump(target_scaler, file_depth + training_data_path + f"{data_group}{folder}25m_added_imdb_context_target_scaler.pkl")

for column in categorical_columns:
    try:
        if label_encoders[column] != {}:
            dump(label_encoders[column], file_depth + training_data_path + f"{data_group}{folder}{column}_label_encoder.pkl")
    except KeyError:
        pass
    

np.save(file_depth + training_data_path + f"{data_group}{folder}train_data.npy", train_data)
np.save(file_depth + training_data_path + f"{data_group}{folder}val_data.npy", val_data)
np.save(file_depth + training_data_path + f"{data_group}{folder}test_data.npy", test_data)
np.save(file_depth + training_data_path + f"{data_group}{folder}train_target_ratings.npy", train_target_ratings)
np.save(file_depth + training_data_path + f"{data_group}{folder}val_target_ratings.npy", val_target_ratings)
np.save(file_depth + training_data_path + f"{data_group}{folder}test_target_ratings.npy", test_target_ratings)
